+++ transcriptic/cli.py
@@ -520,36 +520,37 @@ def resources(ctx, query):
 def inventory(ctx, query):
     """Search organization for inventory"""
     inventory_req = ctx.obj.api.inventory(query)
-    import pdb
-    pdb.set_trace()
     if inventory_req["results"]:
-        kit_req = ctx.obj.api.kits(query)
-        flat_items = list(flatmap(lambda x: [{"name": y["resource"]["name"],
-                                              "id": y["resource"]["id"],
-                                              "vendor": x["vendor"]["name"] if "vendor" in list(x.keys()) else ''}
-                                             for y in x["kit_items"] if
-                                             (y["provisionable"] and not y["reservable"])],
-                                  kit_req["results"]))
-        rs_id_list = [rs["id"] for rs in resource_req["results"]]
-
-        matched_resources = []
-        for item in flat_items:
-            if item["id"] in rs_id_list and item not in matched_resources:
-                matched_resources.append(item)
-
-        if matched_resources:
-            click.echo("Results for '%s':" % query)
-            click.echo('{:^40}'.format("Resource Name") + '|' +
-                       '{:^40}'.format("Vendor") + '|' +
-                       '{:^40}'.format("Resource ID"))
-            click.echo('{:-^120}'.format(''))
-            for resource in matched_resources:
-                click.echo('{:^40}'.format(ascii_encode(resource["name"])) + '|' +
-                           '{:^40}'.format(ascii_encode(resource["vendor"])) + '|' +
-                           '{:^40}'.format(ascii_encode(resource["id"])))
-            click.echo('{:-^120}'.format(''))
-        else:
-            click.echo("No usable resource for '{}'.".format(query))
+        click.echo("Results for '%s':" % query)
+        click.echo('{:^30}'.format("Name") + '|' +
+                   '{:^30}'.format("ID") + '|' +
+                   '{:^30}'.format("Barcode") + '|' +
+                   '{:^30}'.format("Type") + '|' +
+                   '{:^30}'.format("Storage") + '|' +
+                   '{:^30}'.format("Status")
+                   )
+        click.echo('{:-^180}'.format(''))
+        for container in inventory_req["results"]:
+            if container.get("label", None):
+                click.echo('{:^30}'.format(ascii_encode(container["label"])) + '|' +
+                           '{:^30}'.format(ascii_encode(container["id"])) + '|' +
+                           '{:^30}'.format(ascii_encode(container["barcode"])) + '|' +
+                           '{:^30}'.format(ascii_encode(container["container_type_id"])) + '|' +
+                           '{:^30}'.format(ascii_encode(container["storage_condition"])) + '|' +
+                           '{:^30}'.format(ascii_encode(container["status"]))
+                           )
+            # elif container.get("container", None):
+            #     click.echo(
+            #         '{:^30}'.format(ascii_encode(container["container"]["label"])) + '|' +
+            #         '{:^30}'.format(ascii_encode(container["container"]["id"])) + '|' +
+            #         '{:^30}'.format(ascii_encode(container["container"]["barcode"])) + '|' +
+            #         '{:^30}'.format(
+            #             ascii_encode(container["container"]["container_type_id"])) + '|' +
+            #         '{:^30}'.format(
+            #             ascii_encode(container["container"]["storage_condition"])) + '|' +
+            #                '{:^30}'.format(ascii_encode(container["container"]["status"]))
+            #         )
+        click.echo('{:-^180}'.format(''))
     else:
         click.echo("No results for '{}'.".format(query))